# season-1/2.vertical-vs-horizontal-scaling/contracts/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Scaling Lab Baseline API",
    "version": "0.1.0",
    "description": "Source-grounded compatibility baseline at 8cad37f. Known defects are recorded, not endorsed. Full canonical contract is Express; comparison implementations declare capabilities separately."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5102"
    }
  ],
  "paths": {
...
    "/wishlist": {
      "get": {
        "operationId": "listWishlist",
        "summary": "listWishlist",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Baseline success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WishlistList"
                }
              }
            }
          },
          "400": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "description": "Baseline: invalid or nonpositive values default to 20; values over 100 clamp."
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ]
      },
      "post": {
        "operationId": "addWishlist",
        "summary": "addWishlist",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Baseline success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wishlist"
                }
              }
            }
          },
          "400": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WishlistInput"
              }
            }
          }
        }
      }
    },
    "/wishlist/{id}": {
      "get": {
        "operationId": "getWishlist",
        "summary": "getWishlist",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Baseline success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wishlist"
                }
              }
            }
          },
          "400": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ]
      },
      "delete": {
        "operationId": "deleteWishlist",
        "summary": "deleteWishlist",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Baseline success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedWishlist"
                }
              }
            }
          },
          "400": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Validation, authentication, missing resource, conflict, or service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Product": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "description",
          "price",
          "stock",
          "category",
          "imageUrl"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "stock": {
            "type": "integer",
            "minimum": 0
          },
          "category": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Wishlist": {
        "type": "object",
        "required": [
          "_id",
          "userId",
          "productId"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$"
          },
          "userId": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$"
          },
          "productId": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-fA-F0-9]{24}$",
                "nullable": true
              },
              {
                "$ref": "#/components/schemas/Product"
              }
            ]
          }
        }
      },
      "WishlistInput": {
        "type": "object",
        "required": [
          "productId"
        ],
        "properties": {
          "productId": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$"
          }
        }
      },
      "WishlistList": {
        "type": "object",
        "required": [
          "items",
          "nextCursor"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Wishlist"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "DeletedWishlist": {
        "type": "object",
        "required": [
          "deleted"
        ],
        "properties": {
          "deleted": {
            "$ref": "#/components/schemas/Wishlist"
          }
        }
      },
...
    }
  }
}